Improved Windows compatibility for machine command
