+++ pkg/machine/qemu/options_windows.go
@@ -1,0 +1,13 @@
+package qemu
+
+import (
+	"os"
+)
+
+func getRuntimeDir() (string, error) {
+	tmpDir, ok := os.LookupEnv("TEMP")
+	if !ok {
+		tmpDir = os.Getenv("LOCALAPPDATA") + "\\Temp"
+	}
+	return tmpDir, nil
+}
